+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LACLI006 - 21/04/2023 - Add the exceptions of connections.
 - LACLI006 - 21/04/2023 - Add the CommunicatorBase class.
 - LACLI006 - 21/04/2023 - Add the AESCommunicator class.
+- LACLI006 - 21/04/2023 - Add the RSACommunicator class.
 
 ### Changed